% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CompositeData-class}
\alias{CompositeData-class}
\title{A S4 class to represent data for creating a composite}
\description{
A S4 class to represent data for creating a composite
}
\section{Slots}{

\describe{
\item{\code{rawdata}}{A data frame of the data to be used for the composite scores}

\item{\code{groups}}{A character string, the same length as the number of rows of
the data indicating to which group each row belong.  May be all the same
if only one group present in the data.}

\item{\code{thresholds}}{A list with as many elements as there are unique
groups in the data, and where each element is a vector the same
length as the number of columns in the data frame, indicating the
reference thresholds for each variable, by group (all the same if only
one group).}

\item{\code{higherisbetter}}{A logical vector the same length as the number of
columns in the data frame, indicating whether higher is better for
each variable (if \code{TRUE}) and otherwise (if \code{FALSE}) that
lower is better, indicating that variable should be reversed.}

\item{\code{k}}{The number of variables as an integer.}

\item{\code{rawtrans}}{A list of functions to transform the raw data (and thresholds).}
}}

