% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CompositeReady-class}
\alias{CompositeReady-class}
\title{An S4 class to represent composite ready data}
\description{
An S4 class to represent composite ready data
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame ready for use to generate composite scores}

\item{\code{covmat}}{A covariance matrix}

\item{\code{sigma}}{The standard deviation of each variable}

\item{\code{standardize}}{A logical value whether standardization was applied}

\item{\code{distances}}{A data frame of the distance scores}

\item{\code{distanceDensity}}{A ggplot2 density graph}

\item{\code{winsorizedValues}}{A data frame of the values at which each variable was winsorized.
If percentile is 0, that means no winsorization, and values will simply be min and max.}

\item{\code{better}}{A logical value whether better scores than threshold were allowed}

\item{\code{use.prethreshold}}{A logical value whether covariance matrix was calculated from
pre threshold but post winsorizing data.}
}}

