% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{DistanceScores-class}
\alias{DistanceScores-class}
\title{An S4 class to represent distance scores}
\description{
An S4 class to represent distance scores
}
\section{Slots}{

\describe{
\item{\code{distances}}{A data frame of the distance scores}

\item{\code{distanceDensity}}{A ggplot2 density graph}

\item{\code{winsorizedValues}}{A data frame of the values at which each variable was winsorized.
If percentile is 0, that means no winsorization, and values will simply be min and max.}

\item{\code{better}}{A logical value whether better scores than threshold were allowed}
}}

