% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/composite.R
\name{distanceScores}
\alias{distanceScores}
\title{Calculate distance scores on data in preparation for composite scoring}
\usage{
distanceScores(object, winsorize = 0, better = TRUE)
}
\arguments{
\item{object}{An object of class \sQuote{CompositeData}.}

\item{winsorize}{Whether to winsorize the data or not.  Defaults to \code{FALSE}.
If not \code{FALSE}, the percentile to winsorize at.  For example, .01 would be
the .01 and the 1 - .01 percentiles.}

\item{better}{Logical indicating whether \dQuote{better} values than the threshold
are allowed. Defaults to \code{TRUE}.}

\item{na.rm}{A logical whether missing values should be ommitted. Defaults to
\code{TRUE}.}

\item{saveall}{A logical whether to save all intermediary datasets and graphs.
Defaults to \code{FALSE}.}
}
\value{
A list of results.
}
\description{
Calculate distance scores on data in preparation for composite scoring
}
\examples{
# this example creates distances for the built in mtcars data
# see ?mtcars for more details
# The distances are calculated from the "best" in the dataset
# First we create an appropriate CompositeData class object
# higher mpg & hp are better and lower wt & qsec are better
d <- new("CompositeData",
  mtcars[, c("mpg", "hp", "wt", "qsec")],
  thresholds = list(one = with(mtcars, c(
    mpg = max(mpg),
    hp = max(hp),
    wt = min(wt),
    qsec = min(qsec)))
  ),
  higherisbetter = c(TRUE, TRUE, FALSE, FALSE))

# create the distance scores
dres <- distanceScores(d)

# see a density plot of the distance scores
dres@density
# regular summary of distance scores
summary(dres@distances)

# cleanup
rm(d, dres)
}
\seealso{
Other composite: \code{\link{factorComposite}};
  \code{\link{mahalanobisComposite}};
  \code{\link{prepareComposite}};
  \code{\link{sumComposite}}
}

