% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{dropData}
\alias{dropData}
\title{Drop unnecessary data from a MahalanobisComposite object.}
\usage{
dropData(object)
}
\arguments{
\item{object}{A MahalanobisComposite object}
}
\value{
A MahalanobisComposite object with some slots replaced with missing values
}
\description{
This function removes graphs and other sensitive data.
}
\examples{
# make me!!!
}

