% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{initialize,CompositeData-method}
\alias{initialize,CompositeData}
\alias{initialize,CompositeData-method}
\title{An S4 initialize method for CompositeData class}
\usage{
\S4method{initialize}{CompositeData}(.Object, data, groups, thresholds,
  higherisbetter)
}
\arguments{
\item{.Object}{an object of class CompositeData}

\item{data}{A data frame with at least one row and column}

\item{groups}{an optional character vector}

\item{thresholds}{an optional list}

\item{higherisbetter}{an optional logical vector}
}
\description{
An S4 initialize method for CompositeData class
}

