% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predictions.R
\name{.preparePredict}
\alias{.preparePredict}
\title{Internal function to prepare data for prediction}
\usage{
.preparePredict(object, newdata, groups)
}
\arguments{
\item{object}{An object of S4 class \dQuote{MahalanobisScores},
\dQuote{SumScores}, or \dQuote{FactorScores} containing
a model and results to be used to get predictions on new data.}

\item{newdata}{A data frame with identical variable names as was used
to build the initial model.}

\item{groups}{A vector with the same length as the data frame in \code{newdata},
has rows, containing the groups each row belongs to.  See \code{CompositeData}
for more details.}
}
\value{
An object of S4 class \dQuote{CompositeReady}
}
\description{
Internal function to prepare data for prediction
}

