% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{CompositeData}
\alias{CompositeData}
\title{An constructor function for the S4 CompositeData class}
\usage{
CompositeData(rawdata, groups, thresholds, higherisbetter, k)
}
\arguments{
\item{rawdata}{A data frame with at least one row and column}

\item{groups}{an optional character vector. If omitted defaults to
a character vector of all \dQuote{one}s.}

\item{thresholds}{an optional named list where names match the names in groups.
If using defaults for groups, should name it \dQuote{a}.}

\item{higherisbetter}{an optional logical vector}

\item{k}{an optional integer, the number of columns in the raw data}
}
\value{
An S4 object of class \dQuote{CompositeData}
}
\description{
An constructor function for the S4 CompositeData class
}
\examples{
#make me!
}

