% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CompositeReady-class}
\alias{CompositeReady-class}
\title{An S4 class to represent composite ready data}
\description{
An S4 class to represent composite ready data
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame ready for use to generate composite scores}

\item{\code{covmat}}{A covariance matrix}

\item{\code{sigma}}{The standard deviation of each variable}

\item{\code{standardize}}{A logical value whether standardization was applied}
}}

