% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{initialize,CompositeData-method}
\alias{initialize,CompositeData}
\alias{initialize,CompositeData-method}
\title{An S4 initialize method for CompositeData class}
\usage{
\S4method{initialize}{CompositeData}(.Object, data, groups, thresholds,
  higherisbetter)
}
\arguments{
\item{.Object}{an object of class CompositeData}

\item{data}{A data frame with at least one row and column}

\item{groups}{an optional character vector. If omitted defaults to
a character vector of all \dQuote{a}s.}

\item{thresholds}{an optional named list where names match the names in groups.
If using defaults for groups, should name it \dQuote{a}.}

\item{higherisbetter}{an optional logical vector}
}
\description{
An S4 initialize method for CompositeData class
}

